+++ packages/babel-plugin-fbt/src/bin/FbtCollector.js
@@ -21,7 +21,7 @@ export type ExternalTransform = (src: string, opts: TransformOptions, filename: 
 
 /*::
 import type {BabelPluginList, BabelPresetList} from '@babel/core';
-import type {Phrase} from '../index';
+import type {BabelPluginFbt, Phrase} from '../index';
 export type CollectorConfig = {|
   auxiliaryTexts: boolean,
   fbtCommonPath?: string,
@@ -43,7 +43,7 @@ export type PackagerPhrase = {|
 export type TransformOptions = {|
   collectFbt?: boolean,
   soureType?: string,
-  fbtBabelPluginPath?: string,
+  fbtModule?: BabelPluginFbt,
   filename?: string,
   extraOptions?: mixed,
   fbtEnumManifest?: FbtEnumManifest,
@@ -95,6 +95,7 @@ class FbtCollector implements IFbtCollector {
       auxiliaryTexts: this._config.auxiliaryTexts,
       reactNativeMode: this._config.reactNativeMode,
       fbtCommonPath: this._config.fbtCommonPath,
+      fbtModule: fbt,
     };
     if (filename != null) {
       options.filename = filename;
@@ -106,7 +107,6 @@ class FbtCollector implements IFbtCollector {
 
     const externalTransform = this._config.transform;
     if (externalTransform) {
-      options.fbtBabelPluginPath = path.join(__dirname, '../..');
       externalTransform(source, options, filename);
     } else {
       const transform = require('@fbtjs/default-collection-transform');