+++ CHANGELOG.md
@@ -53,7 +53,7 @@ List of changes for each released npm package version.
 </details>
 
 - 0.16.5:
-  - [chore] Update devDependency of `gulp-strip-docblock-pragmas` 
+  - [chore] Update babel-plugin-fbt dependency version to `^0.20.x`
 
 - 0.16.4:
   - [chore] Update babel-plugin-fbt dependency version to `^0.19.x`
@@ -134,6 +134,13 @@ List of changes for each released npm package version.
     </summary>
 
   </details>
+
+- 0.20.0:
+  - [refactor] Hang FbtUtil modules off of babel-plugin-fbt
+  - [refactor]! Replace `fbtBabelPluginPath` with actual referenct to `fbtModule` in external transform
+
+- 0.19.2:
+  - [chore] Update devDependency of `gulp-strip-docblock-pragmas` 
 
 - 0.19.1:
   - [chore] Update fb-babel-plugin-utils dependency version
@@ -243,6 +250,9 @@ List of changes for each released npm package version.
 
 ### babel-plugin-fbt-runtime versions
 
+- 0.9.18:
+  - [chore] Updated peer dependency on babel-plugin-fbt
+
 - 0.9.17:
   - [chore] Updated devDependency on fb-babel-plugin-utils
 
@@ -311,6 +321,10 @@ List of changes for each released npm package version.
   -  First commit. A native JavaScript implementation of the Tiger hash Algorithm.
 
 ### default-collection-transform versions
+- 0.0.3:
+  - Update peer dependency on babel-plugin-fbt
+  - Pass `{enum: true}` to '@babel/plugin-syntax-flow'
+
 - 0.0.2:
   - Require Babel plugins directly